Divide commands and streams into subcycles
Gives easier handling of the experiment plotting.

Also the commands ahould not only be parsed twice: Once for reading the file, once for "executing" the commands.
One could execute the commands at once, but then error handling of the commands would happen when the file is open.
Now the only errors that could happen while the file is open, are lines that couldnt be parsed when the format
is wrong.